requirements.txt

Flask
python-dotenv
firebase-admin
pytest
